requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
mysql-connector-python==8.3.0
pandas==2.2.1
openpyxl==3.1.2
reportlab==4.1.0 